Move FMD component to separate module
This fixes the WASM build.
