+++ crates/core/component/shielded-pool/src/component.rs
@@ -2,6 +2,7 @@
 
 mod action_handler;
 mod assets;
+mod fmd;
 mod metrics;
 mod note_manager;
 mod shielded_pool;
@@ -9,6 +10,7 @@ mod transfer;
 
 pub use self::metrics::register_metrics;
 pub use assets::{AssetRegistry, AssetRegistryRead};
+pub use fmd::ClueManager;
 pub use note_manager::NoteManager;
 pub use shielded_pool::{ShieldedPool, StateReadExt, StateWriteExt};
 pub use transfer::Ics20Transfer;
