+++ crates/core/component/shielded-pool/src/lib.rs
@@ -26,7 +26,6 @@ pub mod output;
 pub mod spend;
 
 pub use convert::{ConvertCircuit, ConvertProof, ConvertProofPrivate, ConvertProofPublic};
-pub use fmd::ClueManager;
 pub use nullifier_derivation::{
     NullifierDerivationCircuit, NullifierDerivationProof, NullifierDerivationProofPrivate,
     NullifierDerivationProofPublic,
